pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "rfut"
version = "0.1.0"
[tool.poetry.dependencies]
dicttoxml = "^1.7"
discord = "^1.0"
flask = "^1.1"
lxml = "^4.5"
pyfutebol = "^2.2"
python = "^3.8"
python-dotenv = "^0.12.0"
treys = "^0.1.3"
wikipedia = "^1.4"
[tool.poetry.dev-dependencies]
